# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegionKey = (typeof RegionKey)[keyof typeof RegionKey];

export const RegionKey = {
  sensory_cortex: "sensory_cortex",
  association_cortex: "association_cortex",
  prefrontal_cortex: "prefrontal_cortex",
  hippocampus: "hippocampus",
  cerebellum: "cerebellum",
  motor_cortex: "motor_cortex",
} as const;

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole];

export const AgentRole = {
  researcher: "researcher",
  planner: "planner",
  executor: "executor",
  memory: "memory",
  critic: "critic",
  summarizer: "summarizer",
} as const;

export interface Region {
  key: RegionKey;
  role: AgentRole;
  name: string;
  description: string;
  /** Base URL of an Ollama server (e.g. https://my-ollama.koyeb.app) */
  ollamaUrl: string;
  /** Ollama model tag, e.g. qwen2.5:1.5b-instruct */
  model: string;
  systemPrompt: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  temperature: number;
  enabled: boolean;
}

export interface UpdateRegionInput {
  ollamaUrl?: string;
  model?: string;
  systemPrompt?: string;
  temperature?: number;
  enabled?: boolean;
}

export interface PingResult {
  ok: boolean;
  latencyMs: number;
  modelsAvailable?: string[];
  error?: string;
}

export interface CreateRunInput {
  goal: string;
  /**
   * @minimum 1
   * @maximum 20
   */
  maxIterations?: number;
}

export type RunStatus = (typeof RunStatus)[keyof typeof RunStatus];

export const RunStatus = {
  pending: "pending",
  running: "running",
  succeeded: "succeeded",
  failed: "failed",
  cancelled: "cancelled",
} as const;

export interface RunSummary {
  id: string;
  goal: string;
  status: RunStatus;
  createdAt: string;
  completedAt?: string;
  iterations: number;
  finalAnswer?: string;
  error?: string;
}

export interface BrainMessage {
  id: string;
  runId: string;
  region: RegionKey;
  role: AgentRole;
  content: string;
  iteration: number;
  latencyMs?: number;
  createdAt: string;
}

export interface RunDetail {
  run: RunSummary;
  messages: BrainMessage[];
}

export interface DashboardInsights {
  totalRuns: number;
  succeededRuns: number;
  failedRuns: number;
  runningRuns: number;
  totalMessages: number;
  avgIterations: number;
  recentRuns: RunSummary[];
}

export interface RegionActivity {
  region: RegionKey;
  role: AgentRole;
  messageCount: number;
  avgLatencyMs: number;
}

export type ListRunsParams = {
  limit?: number;
};