# lib/api-zod/src/generated/types/agentRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole];

export const AgentRole = {
  researcher: "researcher",
  planner: "planner",
  executor: "executor",
  memory: "memory",
  critic: "critic",
  summarizer: "summarizer",
} as const;