# lib/api-zod/src/generated/types/createRunInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRunInput {
  goal: string;
  /**
   * @minimum 1
   * @maximum 20
   */
  maxIterations?: number;
}