# lib/api-zod/src/generated/types/dashboardInsights.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */
import type { RunSummary } from "./runSummary";

export interface DashboardInsights {
  totalRuns: number;
  succeededRuns: number;
  failedRuns: number;
  runningRuns: number;
  totalMessages: number;
  avgIterations: number;
  recentRuns: RunSummary[];
}