# lib/api-zod/src/generated/types/region.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRole } from "./agentRole";
import type { RegionKey } from "./regionKey";

export interface Region {
  key: RegionKey;
  role: AgentRole;
  name: string;
  description: string;
  /** Base URL of an Ollama server (e.g. https://my-ollama.koyeb.app) */
  ollamaUrl: string;
  /** Ollama model tag, e.g. qwen2.5:1.5b-instruct */
  model: string;
  systemPrompt: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  temperature: number;
  enabled: boolean;
}