# lib/api-zod/src/generated/types/regionActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRole } from "./agentRole";
import type { RegionKey } from "./regionKey";

export interface RegionActivity {
  region: RegionKey;
  role: AgentRole;
  messageCount: number;
  avgLatencyMs: number;
}