# lib/api-zod/src/generated/types/regionKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */

export type RegionKey = (typeof RegionKey)[keyof typeof RegionKey];

export const RegionKey = {
  sensory_cortex: "sensory_cortex",
  association_cortex: "association_cortex",
  prefrontal_cortex: "prefrontal_cortex",
  hippocampus: "hippocampus",
  cerebellum: "cerebellum",
  motor_cortex: "motor_cortex",
} as const;