# lib/api-zod/src/generated/types/runStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */

export type RunStatus = (typeof RunStatus)[keyof typeof RunStatus];

export const RunStatus = {
  pending: "pending",
  running: "running",
  succeeded: "succeeded",
  failed: "failed",
  cancelled: "cancelled",
} as const;