# lib/api-zod/src/generated/types/updateRegionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuroLinked Brain — multi-agent Ollama orchestration API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRegionInput {
  ollamaUrl?: string;
  model?: string;
  systemPrompt?: string;
  temperature?: number;
  enabled?: boolean;
}